fix(setup): point URL to repo, parse reqs better.
Once open-sourced, one.wiki will not be accessible to the outside
world. By default, URL points at the github repo as the authoritative
source of information.

When parsing requirements in setup.py, os.path.join is required to
handle clean installation on Windows, since '/' isn't recognized as a
path separator.