Relax dependency pinning
This makes curryproxy a bit more flexible with respect to its
dependencies. They're still ancient, but it's an improvement, and the
architecture makes it a bit easier to add support for newer versions.

I'm not a fan of pinning specific versions, for Reasons. But others are,
so this provides a [frozen] extra for people who really need it.