Improve packaging (#53)
* Relax dependency pinning

This makes curryproxy a bit more flexible with respect to its
dependencies. They're still ancient, but it's an improvement, and the
architecture makes it a bit easier to add support for newer versions.

I'm not a fan of pinning specific versions, for Reasons. But others are,
so this provides a [frozen] extra for people who really need it.

* Add wsgi module for use by gunicorn

This is based off the example given in the wiki. It relieves the user of
having to supply their own, provided they want to use CP in its default
configuration.

* Fix the manifest

check-manifest had some complaints. This fixes all of them.

* Make tox test against different dependency versions

By default it tests all officially-supported versions, but other
versions can be manually tested (for example, if you're trying to
add support for a more recent version of something).

* Make travis test everything in envlist

travis and tox-travis try to do their own venv management, overriding
the envlist supplied in tox.ini. This switches them off and lets tox do
everything.

* Add stub requirements.txt

requirements.txt isn't required, provided setup.py is written correctly.
Some tools (and some people) get confused when it isn't there, though.
This adds a stub and an explanation to avoid future weirdness.