Fix wsgi return type error
PEP 333 specifies that wsgi apps should return an iterable that yields
(byte) strings. Curry instead returns a single string. This worked in
Python 2, but only by accident. Iterating over a Python 2 string gets
you a series of single-character strings, which is not what we want but
went unnoticed because it's technically valid.

Iterating over a Python 3 bytes(), on the other hand, gets you a series
of integers, and things explode.