CBU-1439 - Implement ElasticAPM in Curry
Curry does not utilize an OOB supported WSGI framework for automatic
instrumentation. That said, it is required to manually begin and end
transactions. To start with, I have done this in the simplest way I
could find possible. In the near future, we may want to try adding
more spans to the traces collected as well as additional context
information to the transactions, to get more use out of the agent.

Unfortunately, a lot of trace detail NewRelic was able to gather
automatically to report on will not be available at this time.